Moved some of the redundent code into a utility class
